connect-form: break out components
