+++ src/renderer/components/NestedAccordionDetails.tsx
@@ -1,0 +1,6 @@
+import { AccordionDetails, styled } from '@mui/material';
+
+const NestedAccordionDetails = styled(AccordionDetails)(({ theme }) => ({
+  borderTop: `1px solid ${theme.palette.divider}`,
+}));
+export default NestedAccordionDetails;
