+++ src/renderer/components/NestedAccordionSummary.tsx
@@ -1,0 +1,23 @@
+import { AccordionSummary, AccordionSummaryProps, styled } from '@mui/material';
+import React from 'react';
+import { ChevronRight } from 'react-feather';
+
+const NestedAccordionSummary = styled((props: AccordionSummaryProps) => (
+  <AccordionSummary
+    expandIcon={<ChevronRight fontSize="0.9rem" />}
+    {...props}
+  />
+))(({ theme }) => ({
+  backgroundColor:
+    theme.palette.mode === 'dark'
+      ? 'rgba(255, 255, 255, .05)'
+      : 'rgba(0, 0, 0, .03)',
+  flexDirection: 'row-reverse',
+  '& .MuiAccordionSummary-expandIconWrapper.Mui-expanded': {
+    transform: 'rotate(90deg)',
+  },
+  '& .MuiAccordionSummary-content': {
+    marginLeft: theme.spacing(1),
+  },
+}));
+export default NestedAccordionSummary;
