+++ src/renderer/pages/ConnectForm.tsx
@@ -6,24 +6,15 @@ import {
   Autocomplete,
   Button,
   CardContent,
-  Chip,
   Container,
-  FormControlLabel,
-  FormHelperText,
   Grid,
-  IconButton,
-  styled,
-  Switch,
   Typography,
 } from '@mui/material';
-import { AccordionProps } from '@mui/material/Accordion';
-import { AccordionSummaryProps } from '@mui/material/AccordionSummary';
 import React, { FC, useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
-import { CheckCircle, ChevronDown, ChevronRight, Trash } from 'react-feather';
+import { CheckCircle, ChevronDown } from 'react-feather';
 import { ipcRenderer } from 'electron';
 import { useSnackbar } from 'notistack';
-import { set } from 'lodash';
 import {
   GET_RECORDS,
   GET_UNIQUE_TAGS,
@@ -35,79 +26,9 @@ import {
 import TextField from '../components/TextField';
 import StyledCard from '../components/StyledCard';
 import { formatTag } from '../../shared/validators';
-import {
-  ClientCertFromStore,
-  Connection,
-  Record,
-  Selector,
-} from '../../shared/pb/api';
+import { Connection, Record, Selector } from '../../shared/pb/api';
 import BeforeBackActionDialog from '../components/BeforeBackActionDialog';
-import CertDetails from '../components/CertDetails';
-import CertFilter from '../components/CertFilter';
-
-export const TextArea = styled(TextField)({
-  '& div.MuiFilledInput-root': {
-    background: `rgba(110, 67, 232, 0.05)`,
-    padding: `0px`,
-    marginTop: `10px`,
-    display: `flex`,
-    flexFlow: `row nowrap`,
-    boxShadow: `0 0 0 1px rgb(63 63 68 / 5%), 0 1px 2px 0 rgb(63 63 68 / 15%)`,
-
-    '& div.MuiFilledInput-root': {
-      margin: `2px 0px 0px 6px`,
-      height: `100%`,
-    },
-    '& input.MuiInputBase-input': {
-      padding: `6px`,
-      margin: `6px`,
-    },
-
-    '& .MuiFilledInput-inputMultiline': {
-      padding: `10px`,
-    },
-  },
-  '& div.MuiFilledInput-underline:before': {
-    borderBottom: `0px`,
-  },
-  '& div.MuiFilledInput-underline:after': {
-    border: `0px`,
-  },
-});
-
-const NestedAccordion = styled((props: AccordionProps) => (
-  <Accordion disableGutters elevation={0} {...props} />
-))(({ theme }) => ({
-  border: `1px solid ${theme.palette.divider}`,
-  marginLeft: theme.spacing(2),
-  width: '100%',
-  '&:before': {
-    display: 'none',
-  },
-}));
-
-const NestedAccordionSummary = styled((props: AccordionSummaryProps) => (
-  <AccordionSummary
-    expandIcon={<ChevronRight sx={{ fontSize: '0.9rem' }} />}
-    {...props}
-  />
-))(({ theme }) => ({
-  backgroundColor:
-    theme.palette.mode === 'dark'
-      ? 'rgba(255, 255, 255, .05)'
-      : 'rgba(0, 0, 0, .03)',
-  flexDirection: 'row-reverse',
-  '& .MuiAccordionSummary-expandIconWrapper.Mui-expanded': {
-    transform: 'rotate(90deg)',
-  },
-  '& .MuiAccordionSummary-content': {
-    marginLeft: theme.spacing(1),
-  },
-}));
-
-const NestedAccordionDetails = styled(AccordionDetails)(({ theme }) => ({
-  borderTop: `1px solid ${theme.palette.divider}`,
-}));
+import AdvancedConnectionSettings from '../components/AdvancedConnectionSettings';
 
 interface Props {
   onComplete?: () => void;
@@ -124,17 +45,6 @@ const initialConnData: Connection = {
   clientCertFromStore: undefined,
 };
 
-export function getClientCertFiltersSummary(c: ClientCertFromStore): string {
-  const filters = [];
-  if (c?.issuerFilter) {
-    filters.push('Issuer ' + c.issuerFilter);
-  }
-  if (c?.subjectFilter) {
-    filters.push('Subject ' + c.subjectFilter);
-  }
-  return filters.join(', ');
-}
-
 const ConnectForm: FC<Props> = () => {
   const [showBackWarning, setShowBackWarning] = useState(false);
   const [tags, setTags] = useState<string[]>([]);
@@ -146,12 +56,6 @@ const ConnectForm: FC<Props> = () => {
   };
   const { connectionID } = useParams();
   const { enqueueSnackbar } = useSnackbar();
-  const certRef = React.useRef<HTMLInputElement>(null);
-  const keyRef = React.useRef<HTMLInputElement>(null);
-  const [certText, setCertText] = useState('');
-  const [keyText, setKeyText] = useState('');
-  const [showCertInput, setShowCertInput] = useState(false);
-  const [showDetail, setShowDetail] = useState(false);
 
   useEffect(() => {
     ipcRenderer.on(GET_RECORDS, (_, args) => {
@@ -165,7 +69,6 @@ const ConnectForm: FC<Props> = () => {
         const { conn } = args.res.records[0];
         setConnection(conn || initialConnData);
         setOriginalConnection(conn || initialConnData);
-        setShowCertInput(!conn.clientCert);
       }
     });
     ipcRenderer.once(GET_UNIQUE_TAGS, (_, args) => {
@@ -181,8 +84,6 @@ const ConnectForm: FC<Props> = () => {
         ids: [connectionID],
         tags: [],
       } as Selector);
-    } else {
-      setShowCertInput(true);
     }
     return function cleanup() {
       ipcRenderer.removeAllListeners(GET_RECORDS);
@@ -213,101 +114,12 @@ const ConnectForm: FC<Props> = () => {
     });
   };
 
-  const savePomeriumUrl = (value: string): void => {
-    setConnection({
-      ...connection,
-      ...{ pomeriumUrl: value.trim() || undefined },
-    });
-  };
-
-  const saveDisableTLS = (): void => {
-    setConnection({
-      ...connection,
-      ...{ disableTlsVerification: !connection?.disableTlsVerification },
-    });
-  };
-
   const handleClickBack = (): void => {
     if (JSON.stringify(originalConnection) !== JSON.stringify(connection)) {
       setShowBackWarning(true);
     } else {
       ipcRenderer.send(VIEW_CONNECTION_LIST);
     }
-  };
-
-  const [clientCertFiltersExpanded, setClientCertFiltersExpanded] =
-    useState(false);
-
-  const saveClientCertFromStore = (
-    value: ClientCertFromStore | undefined,
-  ): void => {
-    setConnection({
-      ...connection,
-      ...{ clientCertFromStore: value },
-    });
-    if (value === undefined) {
-      setClientCertFiltersExpanded(false);
-    }
-  };
-
-  const saveClientCertIssuerFilter = (value: string | undefined): void => {
-    saveClientCertFromStore({
-      ...connection?.clientCertFromStore,
-      ...{ issuerFilter: value },
-    });
-  };
-
-  const saveClientCertSubjectFilter = (value: string | undefined): void => {
-    saveClientCertFromStore({
-      ...connection?.clientCertFromStore,
-      ...{ subjectFilter: value },
-    });
-  };
-
-  const clientCertFiltersSummary = getClientCertFiltersSummary(
-    connection?.clientCertFromStore,
-  );
-
-  const saveCertText = (value: string): void => {
-    setCertText(value);
-    setConnection((oldConnection) => {
-      set(oldConnection, 'clientCert.cert', new TextEncoder().encode(value));
-      return oldConnection;
-    });
-  };
-
-  const saveKeyText = (value: string): void => {
-    setKeyText(value);
-    setConnection((oldConnection) => {
-      set(oldConnection, 'clientCert.key', new TextEncoder().encode(value));
-      return oldConnection;
-    });
-  };
-
-  const handleCertFile = (evt) => {
-    const file = evt.target.files[0];
-    const reader = new FileReader();
-    reader.onload = (e) => {
-      saveCertText(e?.target?.result as string);
-    };
-    reader.readAsText(file);
-  };
-
-  const handleKeyFile = (evt) => {
-    const file = evt.target.files[0];
-    const reader = new FileReader();
-    reader.onload = (e) => {
-      saveKeyText(e?.target?.result as string);
-    };
-    reader.readAsText(file);
-  };
-
-  const handleDeleteCert = () => {
-    setConnection((oldConnection) => {
-      oldConnection.clientCert = undefined;
-      return oldConnection;
-    });
-    setShowCertInput(true);
   };
 
   const saveConnection = (): void => {
@@ -333,192 +145,14 @@ const ConnectForm: FC<Props> = () => {
   };
 
   const canSave = (): boolean => {
-    if (certText && !keyText) {
+    if (connection?.clientCert?.cert && !connection?.clientCert?.key) {
       return false;
     }
-    if (!!keyText && !certText) {
+    if (!!connection?.clientCert?.key && !connection?.clientCert?.cert) {
       return false;
     }
     return !!connection?.name && !!connection?.remoteAddr.trim();
   };
-
-  const clientCertFromStoreEnabled =
-    connection?.clientCertFromStore !== undefined;
-
-  const manualClientCertSection = (
-    <>
-      {showCertInput && (
-        <Grid item xs={12}>
-          <label htmlFor="cert-file">
-            <input
-              style={{ display: 'none' }}
-              id="cert-file"
-              ref={certRef}
-              type="file"
-              onChange={handleCertFile}
-            />
-            <Button variant="contained" color="primary" component="span">
-              Client Certificate from File
-            </Button>
-          </label>
-        </Grid>
-      )}
-      {showCertInput && (
-        <Grid item xs={12}>
-          <Typography variant="body2">Client Certificate Text</Typography>
-          <TextArea
-            fullWidth
-            variant="filled"
-            value={certText}
-            multiline
-            required={!!keyText}
-            rows={5}
-            placeholder="e.g. copy/paste the cert in PEM format"
-            onChange={(evt): void => saveCertText(evt.target.value)}
-            spellCheck={false}
-          />
-          <FormHelperText sx={{ pl: 2 }}>
-            Add a Client Certificate with the File Selector or Copy/Paste to the
-            Text Area. Key is required if the Certificate is present.
-          </FormHelperText>
-        </Grid>
-      )}
-      {showCertInput && (
-        <Grid item xs={12}>
-          <label htmlFor="key-file">
-            <input
-              style={{ display: 'none' }}
-              id="key-file"
-              ref={keyRef}
-              type="file"
-              onChange={handleKeyFile}
-            />
-            <Button variant="contained" color="primary" component="span">
-              Client Certificate Key from File
-            </Button>
-          </label>
-        </Grid>
-      )}
-      {showCertInput && (
-        <Grid item xs={12}>
-          <Typography variant="body2">Client Certificate Key Text</Typography>
-          <TextArea
-            fullWidth
-            variant="filled"
-            value={keyText}
-            required={!!certText}
-            multiline
-            rows={5}
-            placeholder="e.g. copy/paste the key in PEM format"
-            onChange={(evt): void => saveKeyText(evt.target.value)}
-          />
-          <FormHelperText sx={{ pl: 2 }}>
-            Add a Client Certificate Key with the File Selector or Copy/Paste to
-            the Text Area. Certificate is required if the Key is present.
-          </FormHelperText>
-        </Grid>
-      )}
-
-      {!showCertInput && (
-        <Grid item xs={12}>
-          <CertDetails
-            open={showDetail}
-            onClose={() => setShowDetail(false)}
-            certInfo={connection?.clientCert?.info}
-          />
-          <Typography variant="body2">Client Certificate</Typography>
-          <Chip
-            label="Details"
-            color="primary"
-            onClick={() => setShowDetail(true)}
-          />
-          <IconButton
-            aria-label="delete"
-            onClick={handleDeleteCert}
-            color="primary"
-            size="large"
-          >
-            <Trash />
-          </IconButton>
-        </Grid>
-      )}
-    </>
-  );
-
-  let clientCertSection = manualClientCertSection;
-  const supportsClientCertFromStore =
-    process.platform === 'win32' || process.platform === 'darwin';
-  if (supportsClientCertFromStore) {
-    clientCertSection = (
-      <>
-        <Grid item xs={12}>
-          <FormControlLabel
-            control={
-              <Switch
-                checked={clientCertFromStoreEnabled}
-                color="primary"
-                onChange={(evt): void =>
-                  saveClientCertFromStore(evt.target.checked ? {} : undefined)
-                }
-              />
-            }
-            label="Search OS certificate store"
-          />
-          <FormHelperText sx={{ pl: 2 }}>
-            Searches for a client certificate based on the trusted CA names
-            provided in the TLS connection handshake.
-          </FormHelperText>
-        </Grid>
-        <NestedAccordion
-          sx={{ mt: 2 }}
-          disabled={!clientCertFromStoreEnabled}
-          expanded={clientCertFiltersExpanded}
-          onChange={(evt, expanded) => setClientCertFiltersExpanded(expanded)}
-        >
-          <NestedAccordionSummary>
-            <Typography>
-              Additional OS certificate store filters
-              {!clientCertFiltersExpanded && clientCertFiltersSummary && (
-                <>
-                  :<br />
-                  {clientCertFiltersSummary}
-                </>
-              )}
-            </Typography>
-          </NestedAccordionSummary>
-          <NestedAccordionDetails>
-            {clientCertFromStoreEnabled && (
-              <>
-                <CertFilter
-                  label="Issuer Name"
-                  data={connection?.clientCertFromStore?.issuerFilter}
-                  onChange={saveClientCertIssuerFilter}
-                  disabled={!clientCertFromStoreEnabled}
-                />
-                <CertFilter
-                  label="Subject Name"
-                  data={connection?.clientCertFromStore?.subjectFilter}
-                  onChange={saveClientCertSubjectFilter}
-                  disabled={!clientCertFromStoreEnabled}
-                />
-              </>
-            )}
-          </NestedAccordionDetails>
-        </NestedAccordion>
-
-        <NestedAccordion sx={{ my: 2 }}>
-          <NestedAccordionSummary>
-            <Typography>Set client certificate manually</Typography>
-          </NestedAccordionSummary>
-          <NestedAccordionDetails>
-            <Grid container spacing={2} sx={{ pt: 1 }}>
-              {manualClientCertSection}
-            </Grid>
-          </NestedAccordionDetails>
-        </NestedAccordion>
-      </>
-    );
-  }
 
   return (
     <Container maxWidth={false}>
@@ -624,42 +258,10 @@ const ConnectForm: FC<Props> = () => {
             <Typography variant="h5">Advanced Settings</Typography>
           </AccordionSummary>
           <AccordionDetails>
-            <Grid container spacing={2}>
-              <Grid item xs={12}>
-                <TextField
-                  fullWidth
-                  required
-                  label="Pomerium URL"
-                  value={connection?.pomeriumUrl || ''}
-                  onChange={(evt): void => savePomeriumUrl(evt.target.value)}
-                  variant="outlined"
-                  autoFocus
-                  helperText="URL of a Bastion host to use for the initial TLS connection"
-                />
-              </Grid>
-              <Grid item xs={12}>
-                <FormControlLabel
-                  control={
-                    <Switch
-                      checked={connection?.disableTlsVerification}
-                      name="disable-tls-verification"
-                      color="primary"
-                      onChange={saveDisableTLS}
-                    />
-                  }
-                  label="Disable TLS Verification"
-                />
-                <FormHelperText sx={{ pl: 2 }}>
-                  Skips TLS verification. No Cert Authority Needed.
-                </FormHelperText>
-              </Grid>
-              <Grid item xs={12}>
-                <Typography sx={{ fontWeight: 500, pt: 1 }}>
-                  Client certificates
-                </Typography>
-              </Grid>
-              {clientCertSection}
-            </Grid>
+            <AdvancedConnectionSettings
+              connection={connection}
+              onChangeConnection={setConnection}
+            />
           </AccordionDetails>
         </Accordion>
